chore: moved queries into separate files
