+++ src/actions/subscriber-subscribe.ts
@@ -2,9 +2,12 @@
 import { v4 as uuidv4 } from 'uuid'
 import { z } from 'zod'
 import { sendConfirmationEmail } from '../utils/resend/email-sendConfirmation'
-import { createValidateEmailLink, createConfirmEmailLink } from '@/lib/utils'
-import { createSubscriber, getOneSubscriberByEmail } from '@/lib/queries'
+import { createValidateEmailLink } from '@/lib/utils'
 import { Status } from '@/types/Types'
+import {
+  createSubscriber,
+  getOneSubscriberByEmail,
+} from '@/queries/subscribers'
 
 const subscribeSchema = z.object({
   email: z.string().email(),
@@ -14,7 +17,6 @@ export const subscribe = async (formData: FormData) => {
   try {
     const token = uuidv4()
 
-    // parse the email
     const parsed = subscribeSchema.safeParse({ email: formData.get('email') })
     if (!parsed.success) return { error: parsed.error.message }
     const checkedEmail = parsed.data.email.toLowerCase()
@@ -39,10 +41,8 @@ export const subscribe = async (formData: FormData) => {
       status: Status.UNSUBSCRIBED,
     }
 
-    // add new subscriber to the database
     await createSubscriber(newSubscriber)
 
-    // send confirmation email using resend
     const validateEmailLink = createValidateEmailLink(token)
     const { error } = await sendConfirmationEmail(
       checkedEmail,
