+++ src/components/deals/ApprovedDeals.tsx
@@ -1,6 +1,7 @@
-import { getApprovedAdminDeals, getApprovedDeals } from '@/lib/queries'
 import React from 'react'
 import DealsList from './DealsList'
+import { getAdminDeals } from '@/queries/adminDeals'
+import { getDeals } from '@/queries/deals'
 
 interface ApprovedAdminDealsProps {
   limit?: number
@@ -8,6 +9,6 @@ interface ApprovedAdminDealsProps {
 export default async function ApprovedAdminDeals({
   limit,
 }: ApprovedAdminDealsProps) {
-  const deals = await getApprovedAdminDeals(limit)
+  const deals = await getDeals({ limit })
   return <DealsList deals={deals} />
 }
