+++ src/components/deals/ApprovedDealsByCategory.tsx
@@ -1,13 +1,13 @@
-import { getApprovedDealsByCategory } from '@/lib/queries'
 import React from 'react'
 import DealsList from './DealsList'
 import { Category } from '@/types/Types'
+import { getDeals } from '@/queries/deals'
 
 export default async function ApprovedDealsByCategory({
   category,
 }: {
   category: Category
 }) {
-  const deals = await getApprovedDealsByCategory(category)
+  const deals = await getDeals({ category })
   return <DealsList deals={deals} />
 }
